# backend/src/factors/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Factors usage doc for golang api server.",
        "title": "Factors Backend Api",
        "contact": {},
        "license": {},
        "version": "1.0"
    },
    "host": "factors-dev.com:8080",
    "basePath": "/projects",
    "paths": {
...
        "/{project_id}/factor": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Factors"
                ],
                "summary": "To run factors model for the given query.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Project ID",
                        "name": "project_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "json",
                        "description": "Factors query",
                        "name": "query",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/pattern_service_wrapper.FactorGraphResults"
                            }
                        }
                    }
                }
            }
        },
...
        "/{project_id}/query": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CoreQuery"
                ],
                "summary": "To run a particular query from core query or dashboards.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Project ID",
                        "name": "project_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Dashboard ID",
                        "name": "dashboard_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Dashboard Unit ID",
                        "name": "dashboard_unit_id",
                        "in": "query"
                    },
                    {
                        "description": "Query payload",
                        "name": "query",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/handler.QueryRequestPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"result\": model.QueryResult, \"cache\": false, \"refreshed_at\": timestamp}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/{project_id}/v1/query": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "V1Api"
                ],
                "summary": "To run events core query as a query group for user and event count.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Project ID",
                        "name": "project_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Dashboard ID",
                        "name": "dashboard_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Dashboard Unit ID",
                        "name": "dashboard_unit_id",
                        "in": "query"
                    },
                    {
                        "description": "Query group",
                        "name": "query_group",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.QueryGroup"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"result\": model.QueryResult, \"cache\": false, \"refreshed_at\": timestamp}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "handler.QueryRequestPayload": {
            "type": "object",
            "properties": {
                "query": {
                    "type": "string"
                }
            }
        },
...
        "model.Query": {
            "type": "object",
            "properties": {
                "cl": {
                    "type": "string"
                },
                "ec": {
                    "description": "all or any",
                    "type": "string"
                },
                "ewp": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.QueryEventWithProperties"
                    }
                },
                "fr": {
                    "type": "integer"
                },
                "gbp": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.QueryGroupByProperty"
                    }
                },
                "gbt": {
                    "type": "object"
                },
                "ovp": {
                    "description": "Deprecated: Keeping it for old dashboard units.",
                    "type": "boolean"
                },
                "see": {
                    "type": "integer"
                },
                "sse": {
                    "type": "integer"
                },
                "to": {
                    "type": "integer"
                },
                "ty": {
                    "type": "string"
                },
                "tz": {
                    "type": "string"
                }
            }
        },
        "model.QueryEventWithProperties": {
            "type": "object",
            "properties": {
                "na": {
                    "type": "string"
                },
                "pr": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.QueryProperty"
                    }
                }
            }
        },
        "model.QueryGroup": {
            "type": "object",
            "properties": {
                "query_group": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Query"
                    }
                }
            }
        },
        "model.QueryGroupByProperty": {
            "type": "object",
            "properties": {
                "en": {
                    "description": "Entity: user or event.",
                    "type": "string"
                },
                "ena": {
                    "description": "group by specific event name.",
                    "type": "string"
                },
                "eni": {
                    "type": "integer"
                },
                "grn": {
                    "description": "currently used only for datetime - year/month/week/day/hour",
                    "type": "string"
                },
                "in": {
                    "type": "integer"
                },
                "pr": {
                    "type": "string"
                },
                "pty": {
                    "description": "Property type categorical / numerical.",
                    "type": "string"
                }
            }
        },
        "model.QueryProperty": {
            "type": "object",
            "properties": {
                "en": {
                    "description": "Entity: user or event.",
                    "type": "string"
                },
                "lop": {
                    "type": "string"
                },
                "op": {
                    "type": "string"
                },
                "pr": {
                    "type": "string"
                },
                "ty": {
                    "description": "Type: categorical or numerical",
                    "type": "string"
                },
                "va": {
                    "type": "string"
                }
            }
        },
...
        "pattern_service_wrapper.FactorGraphResults": {
            "type": "object",
            "properties": {
                "charts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pattern_service_wrapper.graphResult"
                    }
                }
            }
        },
        "pattern_service_wrapper.graphResult": {
            "type": "object",
            "properties": {
                "datasets": {
                    "type": "array",
                    "items": {
                        "type": "\u0026{%!s(token.Pos=1941) string %!s(*ast.InterfaceType=\u0026{1952 0xc0023ed050 false})}"
                    }
                },
                "explanations": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "header": {
                    "type": "string"
                },
                "labels": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "type": {
                    "type": "string"
                },
                "x_label": {
                    "type": "string"
                },
                "y_label": {
                    "type": "string"
                }
            }
        }
    }
}